src/spoke1/admin/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymysql==1.1.1
azure-storage-blob==12.24.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
